Ya esta terminado el modelo con pruebas y demás.

+++ src/main/java/ar/com/jtierno/solarsystem/model/Planet.java
@@ -26,7 +26,7 @@ public class Planet {
         return sunDistance;
     }
 
-    public Vector predictPosition(Integer day){
+    public Vector predictPosition(Long day){
         Double angularPosition = day * angularSpeed;
         Double xPos = Math.cos(angularPosition) * sunDistance;
         Double yPos = Math.sin(angularPosition) * sunDistance;
